Move makeMergedSchema.
