+++ packages/api/src/main.ts
@@ -1,4 +1,4 @@
 export * from 'apollo-server-lambda'
 
-export { makeMergedSchema } from './makeMergedSchema'
+export { makeMergedSchema } from './makeMergedSchema/makeMergedSchema'
 export { server } from './server'
